Consolidating duplicate func test helper method and fixing typos.
