+++ tests/functional/spec/list_collector.spec.js
@@ -1,4 +1,4 @@
-import { checkPeopleInList } from "../helpers";
+import { checkItemsInList } from "../helpers";
 import AnotherListCollectorPage from "../generated_pages/list_collector/another-list-collector-block.page.js";
 import AnotherListCollectorAddPage from "../generated_pages/list_collector/another-list-collector-block-add.page.js";
 import AnotherListCollectorEditPage from "../generated_pages/list_collector/another-list-collector-block-edit.page.js";
@@ -51,7 +51,7 @@ describe("List Collector", () => {
 
     it("The collector shows all of the household members in the summary", async () => {
       const peopleExpected = ["Marcus Twin", "Samuel Clemens", "Olivia Clemens", "Suzy Clemens"];
-      checkPeopleInList(peopleExpected, ListCollectorPage.listLabel);
+      checkItemsInList(peopleExpected, ListCollectorPage.listLabel);
     });
 
     it("The questionnaire allows the name of a person to be changed", async () => {
@@ -109,7 +109,7 @@ describe("List Collector", () => {
 
     it("The collector shows everyone on the summary", async () => {
       const peopleExpected = ["Samuel Clemens", "Olivia Clemens", "Suzy Clemens", "Clara Clemens", "Jean Clemens"];
-      checkPeopleInList(peopleExpected, ListCollectorPage.listLabel);
+      checkItemsInList(peopleExpected, ListCollectorPage.listLabel);
     });
 
     it("When No is answered on the list collector the user sees an interstitial", async () => {
@@ -125,7 +125,7 @@ describe("List Collector", () => {
 
     it("The collector still shows the same list of people on the summary", async () => {
       const peopleExpected = ["Samuel Clemens", "Olivia Clemens", "Suzy Clemens", "Clara Clemens", "Jean Clemens"];
-      checkPeopleInList(peopleExpected, ListCollectorPage.listLabel);
+      checkItemsInList(peopleExpected, ListCollectorPage.listLabel);
     });
 
     it("The collector allows the user to add another person to the same list", async () => {
